Conditionally show different emoji in notification depending on success/failure
+ moved thumbnail generation out of notificationHelper
